bugfix: mapping_fix 好像还是存在问题, 应该是修复了

+++ mapping_fix.py
@@ -23,6 +23,7 @@ if __name__ == "__main__":
     cfg = Configer()
     ssd = SSD(cfg.PATH, (64 << 30), 8, (4 << 10))
     ssd._mapping._init_mapping()
-    ssd._mapping._load_mapping()
-    os.remove(os.path.join(cfg.PATH, cfg.DICT))
+    ssd._mapping.close = lambda: None
+    if os.path.exists(os.path.join(cfg.PATH, cfg.DICT)):
+        os.remove(os.path.join(cfg.PATH, cfg.DICT))
     ssd.close()
